backend/requirements.txt
Django
django-filter==21.1
djangorestframework==3.13.1
djoser==2.1.0
python-decouple==3.5
drf-extra-fields==3.2.1
gunicorn==20.1.0
Pillow
psycopg2-binary